rename to static_image.dart
