+++ lib/src/static_image.dart
